requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

python=3.7.11
pytorch=1.9.1
pyyaml=6.0
pandas=1.3.5
opencv=3.4.2
tqdm=4.64.1
pickle5=0.0.12
easydict=1.10
segmentation-models-pytorch==0.1.0    
